+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2026-02-07
+
+### Changed
+
+- Performance improvements.
+- Retrogen is disabled by default on structure, biome, and feature filters, unless specifically enabled with
+  `"retrogen": "true"`.
+
+### Fixed
+
+- Fixed structure filters not working properly with multiple rules.
+
 ## [1.0.2] - 2026-02-04
 
 ### Added